CC-26089: Adjust code after cr
